Add Relations to dot output
I have some uncommitted changes trying to process Views. However, these
will be extensive since we currently do not keep either the sizes
provided to the view() command, or the AnalyzeViewResult object.
Furthermore, the transform types in AnalyzeViewResult are not currently
exposed, so it will require a few commits to make that accessible. I am
considering other options.